Removing test files and moving updated resources to terraform providers directory; Next step will be to replace xml generated go code in the main for efficiency and better practice
