--- files_for_testing/config.go
@@ -1,41 +1,0 @@
-// Copyright (c) 2017-2022, Juniper Networks Inc. All rights reserved.
-//
-// License: Apache 2.0
-//
-// THIS SOFTWARE IS PROVIDED BY Juniper Networks, Inc. ''AS IS'' AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-// WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL Juniper Networks, Inc. BE LIABLE FOR ANY
-// DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-
-package main
-
-import (
-	"github.com/Juniper/junos-terraform/terraform_providers/netconf"
-)
-
-// Config is the configuration structure used to instantiate the Netconf provider.
-type Config struct {
-	Host     string
-	Port     int
-	Username string
-	Password string
-	SSHKey   string
-}
-
-// Client returns a new client for the provider to use
-func (c *Config) Client() (netconf.Client, error) {
-	return newClient(c)
-}
-
-func newClient(c *Config) (netconf.Client, error) {
-
-	client, err := netconf.NewClient(c.Username, c.Password, c.SSHKey, c.Host, c.Port)
-	return client, err
-}
