--- files_for_testing/main.go
@@ -1,36 +1,0 @@
-// Copyright (c) 2017-2022, Juniper Networks Inc. All rights reserved.
-//
-// License: Apache 2.0
-//
-// THIS SOFTWARE IS PROVIDED BY Juniper Networks, Inc. ''AS IS'' AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-// WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL Juniper Networks, Inc. BE LIABLE FOR ANY
-// DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-
-package main
-
-import (
-	"context"
-	"log"
-
-	"github.com/hashicorp/terraform-plugin-framework/providerserver"
-)
-
-func main() {
-	ctx := context.Background()
-	opts := providerserver.ServeOpts{
-		Address: "registry.terraform.io/juniper/jtaf",
-	}
-
-	err := providerserver.Serve(ctx, newProvider, opts)
-	if err != nil {
-		log.Fatal(err)
-	}
-}
